fix: use tree shaking and some unsafe directive
