+++ package.json
@@ -38,6 +38,8 @@
 		"watch": "pnpm --parallel --filter '@ts-drp/*' watch"
 	},
 	"dependencies": {
+		"@esbuild-plugins/node-globals-polyfill": "^0.2.3",
+		"@esbuild-plugins/node-modules-polyfill": "^0.2.2",
 		"@types/node": "^22.5.4",
 		"esbuild": "^0.25.1",
 		"pascalcase": "^2.0.0",
